refactor: Move Glific Client Concerns to seperate file
